there is something rotten

+++ objective-c/ip2something.h
@@ -13,7 +13,7 @@
     NSString *folder;
     NSFileHandle *datas;
     NSFileHandle *keys;
-    UInt32 length;
+    NSUInteger length;
 }
 -(id) init;
 -(id) initWithPath:(NSString *) path;
